Moved functionality of the former RightControl Component to PopConfig Component.
The old RightControl Component was deleted.
Work on the Population Chart and corresponding config panel and stylings was done.
Still needs work to set the final build chart button on the config panel to fetch the correct data.

Also set the shell of the Area component.
And lastly moved the PopulationChart's coresponding css files into the same folder.