Starting the process of server api reset.
Changed the sql pool to a regular connection.
Starting with the population data set, I am redoing the apis.

I changed the state on the PopulationChart Component to house two main objects.
One is config the other is data.
This was to make the process more clear when I am fetching data from the databases.
Also switching the endpoints to use routes, models and controller directories.

Last I started to remove the old methods for retrieving data (from index.js (server) ).