Continuing to work with the server endpoints.
Consolidating the data from the database.
Added to the usa_states table to include the totals for population totals from all the counties
for the years 2010-2016.
Set up a default data gathering of the top 10 states so I have some starter data in the components.
Set up the initial render for the pie chart. But it still needs a legend to be added.