Added a native d3 Area.jsx component.
It currently works with the top 10 population data category.
Still needs to render the x-axis with dates rather than index numbers.
Also the color scheme should be adjusted to be clearer as to which states belong to
each data set.
And needs tooltips to hover over each point to show the population.